:sparkles: Base carreras

+++ src/main/java/es/allblue/lizardon/objects/tochikarts/Punto.java
@@ -1,35 +1,35 @@
 package es.allblue.lizardon.objects.tochikarts;
 
 public class Punto {
-    private int x, y, z;
+    private double x, y, z;
 
-    public Punto(int x, int y, int z) {
+    public Punto(double x, double y, double z) {
         this.x = x;
         this.y = y;
         this.z = z;
     }
 
-    public int getX() {
+    public double getX() {
         return x;
     }
 
-    public void setX(int x) {
+    public void setX(double x) {
         this.x = x;
     }
 
-    public int getY() {
+    public double getY() {
         return y;
     }
 
-    public void setY(int y) {
+    public void setY(double y) {
         this.y = y;
     }
 
-    public int getZ() {
+    public double getZ() {
         return z;
     }
 
-    public void setZ(int z) {
+    public void setZ(double z) {
         this.z = z;
     }
